+++ src/main/java/com/fourthread/ozang/module/domain/security/filter/JwtAuthenticationFilter.java
@@ -3,8 +3,8 @@ package com.fourthread.ozang.module.domain.security.filter;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fourthread.ozang.module.common.exception.ErrorDetails;
 import com.fourthread.ozang.module.common.exception.ErrorResponse;
-import com.fourthread.ozang.module.domain.security.SecurityMatchers;
-import com.fourthread.ozang.module.domain.security.UserDetailsImpl;
+import com.fourthread.ozang.module.domain.security.jwt.dto.type.SecurityMatchers;
+import com.fourthread.ozang.module.domain.security.userdetails.UserDetailsImpl;
 import com.fourthread.ozang.module.domain.security.jwt.dto.data.JwtPayloadDto;
 import com.fourthread.ozang.module.domain.security.jwt.JwtService;
 import jakarta.servlet.FilterChain;
