+++ src/main/java/com/fourthread/ozang/module/domain/security/jwt/dto/type/SecurityMatchers.java
@@ -1,4 +1,4 @@
-package com.fourthread.ozang.module.domain.security;
+package com.fourthread.ozang.module.domain.security.jwt.dto.type;
 
 public class SecurityMatchers {
 
@@ -9,6 +9,6 @@ public class SecurityMatchers {
   public static final String SIGN_UP = "/api/users";
   public static final String REFRESH = "/api/auth/refresh";
   public static final String OAUTH2 = "/api/oauth2/**";
-  public static final String RESETPASSWORD = "/api/auth/reset-password";
+  public static final String RESET_PASSWORD = "/api/auth/reset-password";
   public static final String CSRF_TOKEN = "/api/auth/csrf-token";
 }
